saved_model_kendaraan/requirements.txt
streamlit
tensorflow<2.16
pillow
numpy
